# openapi.json
{"openapi":"3.1.0","info":{"title":"Qwen-TTS API","version":"1.0.0"},"paths":{"/api/v1/voice-design":{"post":{"tags":["TTS"],"summary":"Generate Voice Design","description":"Generate audio from text description (Voice Design).\nReturns base64 encoded WAV audio list.","operationId":"generate_voice_design_api_v1_voice_design_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoiceDesignRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TTSResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/v1/custom-voice":{"post":{"tags":["TTS"],"summary":"Generate Custom Voice","description":"Generate audio using a specific speaker (Custom Voice).\nReturns base64 encoded WAV audio list.","operationId":"generate_custom_voice_api_v1_custom_voice_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomVoiceRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TTSResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/v1/voice-clone-file":{"post":{"tags":["TTS"],"summary":"Generate Voice Clone File","description":"Clone a voice from an uploaded reference audio file.\nReturns base64 encoded WAV audio list.","operationId":"generate_voice_clone_file_api_v1_voice_clone_file_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_generate_voice_clone_file_api_v1_voice_clone_file_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TTSResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/v1/voice-clone":{"post":{"tags":["TTS"],"summary":"Generate Voice Clone","description":"Clone a voice from reference audio (path, URL, or File URI).\nReturns base64 encoded WAV audio list with custom IDs if provided.","operationId":"generate_voice_clone_api_v1_voice_clone_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoiceCloneRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TTSResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_generate_voice_clone_file_api_v1_voice_clone_file_post":{"properties":{"text":{"type":"string","title":"Text"},"ref_audio":{"type":"string","format":"binary","title":"Ref Audio"},"ref_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ref Text"},"language":{"$ref":"#/components/schemas/LanguageEnum","default":"auto"},"custom_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Custom Id"}},"type":"object","required":["text","ref_audio"],"title":"Body_generate_voice_clone_file_api_v1_voice_clone_file_post"},"CustomVoiceRequest":{"properties":{"text":{"anyOf":[{"type":"string"},{"items":{"type":"string"},"type":"array"}],"title":"Text"},"speaker":{"anyOf":[{"type":"string"},{"items":{"type":"string"},"type":"array"}],"title":"Speaker"},"language":{"anyOf":[{"$ref":"#/components/schemas/LanguageEnum"},{"items":{"$ref":"#/components/schemas/LanguageEnum"},"type":"array"}],"title":"Language","default":"auto"},"instruct":{"anyOf":[{"type":"string"},{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Instruct"}},"type":"object","required":["text","speaker"],"title":"CustomVoiceRequest","example":{"instruct":"Neutral","language":"en","speaker":["Speaker_001","Speaker_002"],"text":["Hello from speaker A","Hello from speaker B"]}},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LanguageEnum":{"type":"string","enum":["auto","en","zh","ja","ko","fr","de","es","ru","pt","it","nl"],"title":"LanguageEnum"},"TTSResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/TTSResponseItem"},"type":"array","title":"Items"},"performance":{"type":"number","title":"Performance","default":0.0}},"type":"object","required":["items"],"title":"TTSResponse"},"TTSResponseItem":{"properties":{"audio_base64":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Audio Base64"},"url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Url"},"custom_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Custom Id"}},"type":"object","title":"TTSResponseItem"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VoiceCloneRequest":{"properties":{"text":{"anyOf":[{"type":"string"},{"items":{"type":"string"},"type":"array"}],"title":"Text"},"ref_audio":{"anyOf":[{"type":"string"},{"items":{"type":"string"},"type":"array"}],"title":"Ref Audio"},"ref_text":{"anyOf":[{"type":"string"},{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Ref Text"},"language":{"anyOf":[{"$ref":"#/components/schemas/LanguageEnum"},{"items":{"$ref":"#/components/schemas/LanguageEnum"},"type":"array"}],"title":"Language","default":"auto"},"custom_id":{"anyOf":[{"type":"string"},{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Custom Id"}},"type":"object","required":["text","ref_audio"],"title":"VoiceCloneRequest","example":{"custom_id":["id_123","id_456"],"language":"en","ref_audio":["/path/to/audio1.wav","/path/to/audio2.wav"],"ref_text":["Reference text 1",null],"text":["First sentence.","Second sentence."]}},"VoiceDesignRequest":{"properties":{"text":{"anyOf":[{"type":"string"},{"items":{"type":"string"},"type":"array"}],"title":"Text"},"instruct":{"anyOf":[{"type":"string"},{"items":{"type":"string"},"type":"array"}],"title":"Instruct"},"language":{"anyOf":[{"$ref":"#/components/schemas/LanguageEnum"},{"items":{"$ref":"#/components/schemas/LanguageEnum"},"type":"array"}],"title":"Language","default":"auto"}},"type":"object","required":["text","instruct"],"title":"VoiceDesignRequest","example":{"instruct":["Happy","Sad"],"language":["en","en"],"text":["Hello world","This is a batch request"]}}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"x-api-key"}}}}